Feat: Update config, add ratelimiting, introduce finer controls for feature flags

+++ src/util/app/auth.ts
@@ -8,86 +8,107 @@ import {
 	AUTH_COOKIE_NAME,
 	AUTH_HEADER_NAME,
 	BRAND_NAME,
-	COOKIE_SILENT_FAIL,
 	JWT_SECRET,
+	JWT_SECRET_EXPOSED_SILENT_FAIL,
 	SESSION_EXPIRY_SILENT_FAIL,
 	SESSION_ID_SILENT_FAIL,
+	STRICT_SESSION_IP_UA_CHECK,
 	VERIFICATION_TOKEN_EXPIRATION_HOURS,
+	VERIFY_EMAIL,
 	WEB_ENDPOINTS,
 	WEB_ORIGIN,
 	WEB_PROTO,
 } from "@/util/config/auth"
 import db from "@/util/db"
-import { SessionCookieContent } from "@/util/http"
+import { type IRequest, SessionJwtContent } from "@/util/http"
 import { middlewareHandler } from "@/util/http/helpers"
 
-export const authParser = middlewareHandler(async (req, res, next) => {
-	let authCookie: string | string[] | undefined = undefined
+type ParsedJwtData =
+	| {
+			jwtString: string
+			jwtData: SessionJwtContent
+			headerMode: boolean
+	  }
+	| {
+			jwtString: undefined
+			jwtData: undefined
+			headerMode: undefined
+	  }
+
+export function parseJwtFromRequest(req: IRequest): ParsedJwtData {
+	let authString: string | string[] | undefined = undefined
 	let altHeaderMode = false
 	if (req.cookies[AUTH_COOKIE_NAME]) {
-		authCookie = req.cookies[AUTH_COOKIE_NAME]
+		authString = req.cookies[AUTH_COOKIE_NAME]
 	}
-	if (!authCookie && req.headers[AUTH_HEADER_NAME.toLowerCase()]) {
-		authCookie = req.headers[AUTH_HEADER_NAME.toLowerCase()]
+	if (!authString && req.headers[AUTH_HEADER_NAME.toLowerCase()]) {
+		authString = req.headers[AUTH_HEADER_NAME.toLowerCase()]
 		altHeaderMode = true
 	}
-	if (authCookie && typeof authCookie === "string" && authCookie.length > 0) {
-		const parsedCookie = verify(
-			authCookie,
-			JWT_SECRET,
-		) as SessionCookieContent
-		if (!parsedCookie.sessionId) {
-			if (!COOKIE_SILENT_FAIL) {
+
+	if (authString && typeof authString === "string" && authString.length > 0) {
+		const parsedData = verify(authString, JWT_SECRET) as SessionJwtContent
+
+		if (!parsedData || typeof parsedData.sessionId !== "string") {
+			if (!JWT_SECRET_EXPOSED_SILENT_FAIL) {
 				throw new Error(
-					"Session ID was not found in the cookie! The environment secrets may have been leaked!",
+					"Session ID was not found in the jwt! The environment secrets may have been leaked!",
 				)
 			}
-		} else {
-			const dbSession = await db.session.findFirst({
-				where: {
-					sessionId: parsedCookie.sessionId,
-					sessionToken: authCookie,
-				},
-				include: {
-					sessionUser: true,
-					sessionOrg: true,
-				},
-			})
-			if (!dbSession) {
-				if (!SESSION_ID_SILENT_FAIL) {
-					throw new Error(
-						`Session ${parsedCookie.sessionId} not found in the database!`,
-					)
-				}
+			return {
+				jwtString: undefined,
+				jwtData: undefined,
+				headerMode: undefined,
 			}
-
-			if (
-				dbSession &&
-				dbSession.sessionEndTimestamp &&
-				dbSession.sessionEndTimestamp <= new Date()
-			) {
-				if (!SESSION_EXPIRY_SILENT_FAIL) {
-					throw new Error(
-						`Expired / Closed session ${dbSession.sessionId} was used`,
-					)
-				}
-			}
-
-			// Check if the session is from the same IP and User Agent
-			// if (
-			// 	altHeaderMode === false &&
-			// 	dbSession.sessionIp === req.ip &&
-			// 	dbSession.sessionUA === req.headers["user-agent"]
-			// ) {
-			// 	req.currentSession = dbSession
-			// 	req.devApp = undefined
-			// } else if (altHeaderMode === true) {
-			// 	req.currentSession = dbSession
-			// 	req.devApp = undefined
-			// }
-			req.currentSession = dbSession ?? undefined
-			// req.devApp = undefined
 		}
+	}
+
+	return {
+		jwtString: undefined,
+		jwtData: undefined,
+		headerMode: undefined,
+	}
+}
+
+export const authParser = middlewareHandler(async (req, res, next) => {
+	const { headerMode, jwtData, jwtString } = parseJwtFromRequest(req)
+	if (jwtData) {
+		const dbSession = await db.session.findFirst({
+			where: {
+				sessionId: jwtData.sessionId,
+				sessionToken: jwtString,
+			},
+			include: {
+				sessionUser: true,
+				sessionOrg: true,
+			},
+		})
+		if (!dbSession) {
+			if (!SESSION_ID_SILENT_FAIL) {
+				throw new Error(
+					`Session ${jwtData.sessionId} not found in the database!`,
+				)
+			}
+		}
+
+		if (dbSession && dbSession.sessionEndTimestamp <= new Date()) {
+			if (!SESSION_EXPIRY_SILENT_FAIL) {
+				throw new Error(
+					`Expired / Closed session ${dbSession.sessionId} was used`,
+				)
+			}
+			if (STRICT_SESSION_IP_UA_CHECK){
+				const reqIp = req.ip || ""
+				const reqUa = req.headers["user-agent"] || ""	
+				if (
+					dbSession.sessionIp !== reqIp ||
+					dbSession.sessionUA !== reqUa
+				){
+					throw new Error(`Session ${dbSession.sessionId} IP/UA mismatch!`)
+				}
+			}
+		}
+		req.currentSession = dbSession ?? undefined
 	}
 
 	next()
@@ -102,6 +123,9 @@ type SendVerificationTokenArgs = {
 export async function createAndSendVerificationToken(
 	args: SendVerificationTokenArgs,
 ) {
+	if (!VERIFY_EMAIL) {
+		return
+	}
 	// Delete older tokens
 	await db.userVerificationToken.deleteMany({
 		where: {
